Use FS.createLazyFile and optimize builds (#5)
This PR includes two main changes:
1. Using `FS.createLazyFile`.
2. Optimizing builds.

It also bumps to version `2.0.1`.

## `FS.createLazyFile`
The main change in this PR is to use `FS.createLazyFile` for accessing remote files via range requests, rather than patching the openslide code to provide this functionality.

This works well and means we will have an easier time upgrading openslide in the future by not needing to maintain a large patch. There is, however, one possible downside: in addition to the CORS headers already required to access files served by a remote server, we also need to add the following headers to the `Access-Control-Expose-Headers` list: `Accept-Ranges`, `Content-Encoding`, and `Content-Length`. This is because `FS.createLazyFile` checks and respects those headers when performing range requests whereas our current implementation does not.

Ref: https://github.com/emscripten-core/emscripten/blob/4.0.8/src/lib/libfs.js#L1717-L1720

## Optimizing Build
The other change is to optimize the emcc build to get more optimized wasm files:
1. We remove the `-s SINGLE_FILE=1` so that we don't have to embed the wasm as a base64 string and can instead leave it as its own binary `.wasm` file (`.wasm` files being more compressed than the base64 string version).
Ref: https://github.com/emscripten-core/emscripten/blob/4.0.8/src/settings.js#L1847-L1858
4.  We also use the `-Os` option (and remove `-g`) to optimize the WASM (and JS), focusing especially on the size.
Ref: https://emscripten.org/docs/tools_reference/emcc.html#emcc-os

With these 2 optimizations we get from a 31MB `lib.js` to a 102KB `lib.js` + 8.6MB `lib.wasm`. This is without minification and without compression like gzip or brotli (which could be performed by the web server used to serve the files). Note that using `-Os` does take about 4x longer, so in development it can make sense to remove it (as noted in the `build.sh` file).