Fix memory re-use issue in readRegion (#6)
Without calling `.slice()` to get a copy of the data, we were getting a view into memory that was then re-used, so if we called `readRegion` multiple times (in parallel) there often was a race condition where we would get the data from one call passed back in multiple calls.

Also updates `examples/canvas/index.html` to make multiple parallel calls to `readRegion`.