Enable multiple workers (#8)
This provides the configurable option to create N workers when creating the `OpenSlide` instance. This seems to help in cases where we have many `readRegion` calls happening concurrently. The trade-off is that it triggers N network requests for `worker.js`, `lib.js`, and `lib.wasm`, although in local testing it appears that the subsequent calls re-use the initial request's response data.