Build optimizations (#9)
This PR incorporates a bunch of optimizations in the build process (`build.sh`). 

In a test scenario of reading a sample 2.88GB SVS at 10 MPP in a 16x16 grid (i.e. 256 regions to be read, in parallel) with only 1 worker, we go from a total of 17.6s to read all the regions to 7.6s (in Chrome on my laptop).

The main updates are:
* Consistently using "release" build types (`--buildtype=release` for meson builds, `-DCMAKE_BUILD_TYPE=Release` for cmake builds). The main factor in speed-up was in ensuring `openjpeg` was built in release mode.
* Consistently applying cflags and ldflags on each step, including the optimizations `-s O3` for Emscripten optimization (ref: https://emscripten.org/docs/tools_reference/emcc.html#emcc-o3) and `-msimd128` to enable SIMD (ref: https://emscripten.org/docs/porting/simd.html)
* Enable link time optimization through the `-flto` flag (ref: https://emscripten.org/docs/optimizing/Optimizing-Code.html#lto).

It also includes some clean-up of the build scripts (e.g. only needing to pass in cflags and ldflags at setup/compile step, not install step). And we add the `set -e` flag to the build script to avoid silent errors (i.e. script exits on error).

What is not included (yet) is any improvements we might get by using more recent versions of libraries (e.g. openjpeg).