Always return RGBA from readRegion (#14)
Removes the option to not return in RGBA order (which would result in BGRA order) since that is not likely to be used.

For posterity, we considered just changing the default to `true`, but we also don't see a real use case for wanting BGRA order, so for simplicity we are removing that option.

In the future we can clean up the `openslide-api.c` file, but that's not a big deal at this point given it's not in the public facing API.