Implement ability to abort readRegion calls (#15)
This will be helpful when working with tile fetching in a slide viewer where we can abort fetching of tiles that are no longer in view.

In reality it will only "abort" the `readRegion` call if we haven't yet passed into the C layer. When there a lot of concurrent `readRegion` calls (e.g. in a slide viewer, or in the example HTML file in this repo when increasing the number of partitions) this actually does abort a number of calls to `api.readRegion`.

### Usage

```typescript
const controller = new AbortController();
const region = await slide.readRegion(x, y, level, w, h, {
  signal: controller.signal;
});
```
And then somewhere else:
```typescript
controller.abort("Region not needed");
```
At which point `slide.readRegion` would throw an `AbortError` with message `"Region not needed"`.